+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [#6](https://github.com/creekorful/srcode/issues/6) Implement srcode set-cmd.
 
+## Changed
+
+- Improve documentation.
+
 ## [0.2.0] - 2021-01-17
 
 ## Added